requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
gunicorn

firebase-admin~=6.7.0